chore(script-fixes): moved setup-roles to utils, moved punkrouter deployment into deploy.ts
