+++ scripts/bootstrap-state-with-loans.ts
@@ -3,7 +3,6 @@
 import { ethers } from "hardhat";
 
 import { main as deploy } from "./deploy/deploy";
-import { main as deployPunkRouter } from "./deploy-with-punk-router";
 import { SECTION_SEPARATOR, vaultAssetsAndMakeLoans } from "./utils/bootstrap-tools";
 import { mintAndDistribute } from "./utils/mint-distribute-assets";
 import { deployNFTs } from "./utils/deploy-assets";
@@ -19,9 +18,7 @@ export async function main(): Promise<void> {
     console.log("Deploying resources...\n");
 
     // Deploy the smart contracts
-    const { vaultFactory, originationController, borrowerNote, repaymentController } = await deploy();
-    await deployPunkRouter();
-
+    const { vaultFactory, originationController, borrowerNote, repaymentController, punkRouter } = await deploy();
 
     // Mint some NFTs
     console.log(SECTION_SEPARATOR);
